Get rid of internally Buffered components, use setBuffer on all rendering flows
This stands to dramatically reduce memory footprint, particularly on projects with
a large number of pattern instances. In hindsight it was not a good idea to have
every pattern own a distinct copy of an entire pixel buffer, scaling memory usage
on O(N) of patterns is overkill. Instead, we scale buffers O(N) on the number of
mixer channels, which ought to generally be a much lower sane number and allow
for efficient rendering.